requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
torch==2.2.2
torchvision==0.17.2
pillow==10.3.0
python-multipart==0.0.9
